+++ demisto_sdk/commands/common/hook_validations/docker.py
@@ -2,8 +2,9 @@ import re
 from datetime import datetime, timedelta
 from functools import lru_cache
 from typing import Optional, Tuple, Union
-from dateparser import parse
+
 import requests
+from dateparser import parse
 from pkg_resources import parse_version
 
 from demisto_sdk.commands.common.constants import (
@@ -112,59 +113,66 @@ class DockerImageValidator(BaseValidator):
 
     @error_codes("DO100,DO106,DO101")
     def is_docker_image_latest_tag(self):
+        is_latest_tag = True
         if (
             "demisto/python:1.3-alpine"
             == f"{self.docker_image_name}:{self.docker_image_tag}"
         ):
+            self.is_latest_tag = False
             # the docker image is the default one
             error_message, error_code = Errors.default_docker_error()
             if self.handle_error(error_message, error_code, file_path=self.file_path):
-                self.is_latest_tag = False
-
-            return self.is_latest_tag
+                is_latest_tag = False
 
         # ignore tag or non-demisto docker issues
-        if self.docker_image_latest_tag == "no-tag-required":
-            return self.is_latest_tag
+        elif self.docker_image_latest_tag == "no-tag-required":
+            return True
 
-        if not self.docker_image_name or not self.docker_image_latest_tag:
+        elif not self.docker_image_name or not self.docker_image_latest_tag:
             # If the docker image isn't in the format we expect it to be or we failed fetching the tag
             # We don't want to print any error msgs to user because they have already been printed
             # see parse_docker_image for the errors
-            self.is_latest_tag = False
-            return self.is_latest_tag
+            self.is_latest_tag, is_latest_tag = False, False
 
-        if self.docker_image_latest_tag != self.docker_image_tag and self.is_docker_older_than_three_days():
+        elif self.docker_image_latest_tag != self.docker_image_tag:
             # If docker image tag is not the most updated one that exists in docker-hub
-            error_message, error_code = Errors.docker_not_on_the_latest_tag(
-                self.docker_image_tag, self.docker_image_latest_tag, self.is_iron_bank
-            )
-            suggested_fix = Errors.suggest_docker_fix(
-                self.docker_image_name, self.file_path, self.is_iron_bank
-            )
-            if self.handle_error(
-                error_message,
-                error_code,
-                file_path=self.file_path,
-                suggested_fix=suggested_fix,
-            ):
-                self.is_latest_tag = False
-
-            else:
-                # if this error is ignored - do print it as a warning
-                self.handle_error(
-                    error_message, error_code, file_path=self.file_path, warning=True
+            self.is_latest_tag = False
+            if self.is_docker_older_than_three_days():
+                error_message, error_code = Errors.docker_not_on_the_latest_tag(
+                    self.docker_image_tag,
+                    self.docker_image_latest_tag,
+                    self.is_iron_bank,
                 )
+                suggested_fix = Errors.suggest_docker_fix(
+                    self.docker_image_name, self.file_path, self.is_iron_bank
+                )
+                if self.handle_error(
+                    error_message,
+                    error_code,
+                    file_path=self.file_path,
+                    suggested_fix=suggested_fix,
+                ):
+                    is_latest_tag = False
+
+                else:
+                    # if this error is ignored - do print it as a warning
+                    self.handle_error(
+                        error_message,
+                        error_code,
+                        file_path=self.file_path,
+                        warning=True,
+                    )
 
         # the most updated tag should be numeric and not labeled "latest"
         if self.docker_image_latest_tag == "latest":
+            self.is_latest_tag = False
             error_message, error_code = Errors.latest_docker_error(
                 self.docker_image_tag, self.docker_image_name
             )
             if self.handle_error(error_message, error_code, file_path=self.file_path):
-                self.is_latest_tag = False
+                is_latest_tag = False
 
-        return self.is_latest_tag
+        return is_latest_tag
 
     def is_docker_older_than_three_days(self):
         """
@@ -173,9 +181,16 @@ class DockerImageValidator(BaseValidator):
         Returns:
             bool: True if the docker is more than 3 days old.
         """
-        three_days_ago: Optional[datetime] = parse('3 days ago')
-        last_updated = self.get_docker_image_creation_date(self.docker_image_name, self.docker_image_tag)
-        if not last_updated or three_days_ago and three_days_ago > datetime.strptime(last_updated, "%Y-%m-%dT%H:%M:%S.%fZ"):
+        three_days_ago: Optional[datetime] = parse("3 days ago")
+        last_updated = self.get_docker_image_creation_date(
+            self.docker_image_name, self.docker_image_tag
+        )
+        if (
+            not last_updated
+            or three_days_ago
+            and three_days_ago
+            > datetime.strptime(last_updated, "%Y-%m-%dT%H:%M:%S.%fZ")
+        ):
             return True
         return False
 
@@ -306,8 +321,9 @@ class DockerImageValidator(BaseValidator):
 
     @staticmethod
     @lru_cache(256)
-    def get_docker_image_creation_date(docker_image_name: str,
-                                       docker_image_tag: str) -> str:
+    def get_docker_image_creation_date(
+        docker_image_name: str, docker_image_tag: str
+    ) -> str:
         """
         Get the last_updated field of the given docker.
         Args: