+++ demisto_sdk/commands/common/tests/docker_test.py
@@ -269,6 +269,47 @@ class TestDockerImage:
         assert docker_image_validator.is_latest_tag is True
         assert docker_image_validator.is_docker_image_valid() is True
 
+    # disable-secrets-detection-end
+    @pytest.mark.parametrize(
+        "return_value, expected_latest_tag_value, expected_function_results",
+        [("2023-05-19T15:06:52.316769Z", False, False), ("python", True, False)],
+    )
+    def test_not_latest_docker_older_than_3_days(
+        self, mocker, return_value, expected_latest_tag_value, expected_function_results
+    ):
+        """
+        Given
+        - The default docker image - 'demisto/python:1.3-alpine'
+
+        When
+        - The most updated docker image in docker-hub is '1.0.3'
+
+        Then
+        -  Case 1: The current docker image is more than 3 days old and should fail the validation and set is_latest_tag to False.
+        -  Case 2: The current docker image is less than 3 days old and shouldn't fail the validation but set is_latest_tag to False.
+        """
+        docker_image_validator = mock_docker_image_validator()
+        docker_image_validator.docker_image_latest_tag = "1.0.3"
+        docker_image_validator.docker_image_name = "demisto/python"
+        docker_image_validator.code_type = "python"
+        docker_image_validator.is_latest_tag = True
+        docker_image_validator.docker_image_tag = "1.0.2"
+        docker_image_validator.is_valid = True
+        docker_image_validator.yml_docker_image = "demisto/python:1.0.2"
+        docker_image_validator.is_iron_bank = False
+        docker_image_validator.is_deprecated_image = ""
+        mocker.patch.object(
+            docker_image_validator,
+            "get_docker_image_creation_date",
+            return_value=return_value,
+        )
+        assert (
+            docker_image_validator.is_docker_image_latest_tag()
+            is expected_function_results
+        )
+        assert docker_image_validator.is_latest_tag is expected_latest_tag_value
+        assert True
+
     def test_is_docker_image_latest_tag_with_numeric_but_not_most_updated(self):
         """
         Given